Seperate out sampler classes into different modules.
